Enable maintainer mode by default.
Most users tracking the repository will want the autotools-generated
files to be rebuilt if necessary.  At the same time, some distribution
packagers might want to disable this when making packages of releases.

This commit changes the default, so that it is still possible to
disable dependency checking, but autotools will work as expected for
most users.

Signed-off-by: Tim Retout <tim@retout.co.uk>